change: grouping paths from class to method
